# components/db-manager/pkg/api/docs/swagger.yaml
definitions:
...
  models.ErrorResp:
    properties:
      error:
        type: string
    type: object
...
  models.ListResp-models_Media:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Media'
        type: array
      total:
        type: integer
    type: object
...
  models.ListResp-models_Trip:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Trip'
        type: array
      total:
        type: integer
    type: object
...
  models.Media:
    properties:
      ContentType:
        description: ENUM checked at API level
        type: string
      comments:
        items:
          type: string
        type: array
      createdAt:
        type: string
      id:
        type: string
      isAt:
        type: string
      name:
        description: original file name
        type: string
      ownerId:
        type: string
      participants:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      url:
        description: served path
        type: string
    type: object
...
  models.Trip:
    properties:
      createdAt:
        type: string
      descriptionId:
        description: -> comment id
        type: string
      endAt:
        type: string
      id:
        type: string
      name:
        type: string
      ownerId:
        type: string
      participants:
        description: user ids
        items:
          type: string
        type: array
      startAt:
        type: string
      stops:
        description: jsonb
        items:
          $ref: '#/definitions/models.TripStop'
        type: array
      transportation:
        description: enum string
        type: string
      updatedAt:
        type: string
    type: object
  models.TripInputAPI:
    properties:
      descriptionId:
        format: uuid
        type: string
      endAt:
        type: integer
      name:
        example: Euro trip
        type: string
      participants:
        description: UUID strings
        items:
          type: string
        type: array
      startAt:
        type: integer
      stops:
        items:
          $ref: '#/definitions/models.TripStop'
        type: array
      transportation:
        enum:
        - car
        - plane
        - train
        - bus
        - foot
        example: car
        type: string
    type: object
  models.TripStop:
    properties:
      id:
        type: string
      order:
        type: integer
    type: object
...
info:
  contact:
    email: guillermo.gc1994@gmail.com
    name: Guillermo Gomez
    url: https://github.com/guigomcha/famquest/
  description: Handles the connection to the DBs in DB Manager. PostgreSQL and MINIO
  license:
    name: Guillermo Gomez GPL V3
  termsOfService: http://swagger.io/terms/
  title: FamQuest DB Manager API
  version: 0.4.0
paths:
...
  /media:
    get:
      description: Return paginated media list
      parameters:
      - description: page size
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListResp-models_Media'
      summary: List media
      tags:
      - media
    post:
      consumes:
      - multipart/form-data
      description: Multipart upload
      parameters:
      - description: image/* or audio/* or video/ or application/pdf
        in: formData
        name: file
        required: true
        type: file
      - description: unix ms
        in: formData
        name: isAt
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Upload media file
      tags:
      - media
  /media/{id}:
    delete:
      description: Return a single media
      parameters:
      - description: media id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Delete media by id
      tags:
      - media
    get:
      description: Return a single media
      parameters:
      - description: media id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Media'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Get media by id
      tags:
      - media
    put:
      consumes:
      - multipart/form-data
      description: Multipart update
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: image/* or audio/* or video/ or application/pdf
        in: formData
        name: file
        required: true
        type: file
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: comments
        type: array
      - description: unix ms
        in: formData
        name: isAt
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: participants
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Update media file
      tags:
      - media
...
  /trips:
    get:
      description: Return paginated trip list
      parameters:
      - description: page size
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListResp-models_Trip'
      summary: List trips
      tags:
      - trips
    post:
      consumes:
      - application/json
      description: Creates a new trip
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TripInputAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Create trip
      tags:
      - trips
  /trips/{id}:
    delete:
      description: Remove a trip
      parameters:
      - description: trip id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Delete trip
      tags:
      - trips
    get:
      description: Return a single trip
      parameters:
      - description: trip id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trip'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Get trip by id
      tags:
      - trips
    put:
      consumes:
      - application/json
      description: Update an existing trip
      parameters:
      - description: trip id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TripInputAPI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Update trip
      tags:
      - trips
...
swagger: "2.0"